clean up and rename authmiddleware.py
